refactor: 리뷰 상세 페이지 컴포넌트 분리 및 LayoutWrapper 적용
